+++ clang/docs/ReleaseNotes.rst
@@ -54,6 +54,12 @@ ABI Changes in This Version
   ``unsigned char`` to ``uint64_t`` to prevent overflow. This might be an ABI
   break for such structs compared to earlier Clang versions.
 
+- Fixed Microsoft calling convention to match how MSVC returns vector types from
+  C++ member functions on x86/x86-64. This change resolves incompatibilities with
+  code compiled by MSVC but will introduce incompatibilities with code compiled
+  by Clang 21 and earlier versions, unless the ``-fclang-abi-compat=21`` option
+  is used. (#GH104)
+
 AST Dumping Potentially Breaking Changes
 ----------------------------------------